Classification Analysis
In this dataset, we will explore a dataset and perform a classification task using Python. The dataset we will be using is the "Iris" dataset, which contains measurements of different iris flowers along with their species. Our goal is to build a classification model that can predict the species of an iris flower based on its measurements.