Update README.md
This repository contains code for performing classification analysis on the "Classification Analysis" dataset available on Kaggle. The dataset used in this analysis is the "iris.csv" file.

The dataset used in this analysis is the "iris.csv" file. It contains information about various attributes of iris flowers, such as sepal length, sepal width, petal length, petal width, and species. The goal of this analysis is to predict the species of an iris flower based on its attributes.